updated tests, reuse utils for Form and FormWrapper
