+++ packages/vulcan-forms/lib/components/Form.jsx
@@ -61,8 +61,9 @@ import mergeWithComponents from '../modules/mergeWithComponents';
 import {
   getEditableFields,
   getInsertableFields
-} from '../modules/fieldsUtils.js';
+} from '../modules/schema_utils.js';
 import withCollectionProps from './withCollectionProps';
+import { callbackProps } from './propTypes';
 
 const compactParent = (object, path) => {
   const parentPath = getParentPath(path);
@@ -82,8 +83,7 @@ const getDefaultValues = convertedSchema => {
 };
 
 const getInitialStateFromProps = nextProps => {
-  const collection =
-    nextProps.collection || getCollection(nextProps.collectionName);
+  const collection = nextProps.collection;
   const schema = nextProps.schema
     ? new SimpleSchema(nextProps.schema)
     : collection.simpleSchema();
@@ -148,36 +148,14 @@ class SmartForm extends Component {
   // --------------------------------------------------------------------- //
 
   /*
-
-  Get the current collection
-
-  */
-  getCollection = () => {
-    return this.props.collection || getCollection(this.props.collectionName);
-  };
-
-  /*
-
-  Get current typeName
-
-  */
-  getTypeName = () => {
-    return this.getCollection().options.typeName;
-  };
-
-  /*
-
   If a document is being passed, this is an edit form
-
   */
   getFormType = () => {
     return this.props.document ? 'edit' : 'new';
   };
 
   /*
-
   Get a list of all insertable fields
-
   */
   getInsertableFields = schema => {
     return getInsertableFields(
@@ -187,9 +165,7 @@ class SmartForm extends Component {
   };
 
   /*
-
   Get a list of all editable fields
-  
   */
   getEditableFields = schema => {
     return getEditableFields(
@@ -521,7 +497,7 @@ class SmartForm extends Component {
   
    */
   getLabel = (fieldName, fieldLocale) => {
-    const collectionName = this.getCollection().options.collectionName.toLowerCase();
+    const collectionName = this.props.collectionName.toLowerCase();
     const defaultMessage = '|*|*|';
     let id = `${collectionName}.${fieldName}`;
     let intlLabel;
@@ -947,13 +923,13 @@ class SmartForm extends Component {
 
     if (this.getFormType() === 'new') {
       // create document form
-      this.props[`create${this.getTypeName()}`]({ data })
+      this.props[`create${this.props.typeName}`]({ data })
         .then(this.newMutationSuccessCallback)
         .catch(error => this.mutationErrorCallback(document, error));
     } else {
       // update document form
       const documentId = this.getDocument()._id;
-      this.props[`update${this.getTypeName()}`]({
+      this.props[`update${this.props.typeName}`]({
         selector: { documentId },
         data
       })
@@ -999,7 +975,7 @@ class SmartForm extends Component {
 
   render() {
     const fieldGroups = this.getFieldGroups();
-    const collectionName = this.getCollection()._name;
+    const collectionName = this.props.collectionName;
     const FormComponents = mergeWithComponents(this.props.formComponents);
 
     return (
@@ -1062,6 +1038,7 @@ SmartForm.propTypes = {
   collection: PropTypes.object.isRequired,
   collectionName: PropTypes.string.isRequired,
   typeName: PropTypes.string.isRequired,
+
   document: PropTypes.object, // if a document is passed, this will be an edit form
   schema: PropTypes.object, // usually not needed
 
@@ -1086,13 +1063,7 @@ SmartForm.propTypes = {
   formComponents: PropTypes.object,
 
   // callbacks
-  changeCallback: PropTypes.func,
-  submitCallback: PropTypes.func,
-  successCallback: PropTypes.func,
-  removeSuccessCallback: PropTypes.func,
-  errorCallback: PropTypes.func,
-  cancelCallback: PropTypes.func,
-  revertCallback: PropTypes.func,
+  ...callbackProps,
 
   currentUser: PropTypes.object,
   client: PropTypes.object
