+++ packages/vulcan-forms/test/components.test.js
@@ -172,24 +172,24 @@ describe('vulcan-forms/components', function() {
       context
     });
 
-  describe('Form (handle fields computation)', function() {
+  describe('Form collectionName="" (handle fields computation)', function() {
     // getters
     const getArrayFormGroup = wrapper => wrapper.find('FormGroup').find({ name: 'addresses' });
     const getFields = arrayFormGroup => arrayFormGroup.prop('fields');
 
     describe('basic collection - no nesting', function() {
       it('shallow render', function() {
-        const wrapper = shallowWithContext(<Form collection={Addresses} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={Addresses} />);
         expect(wrapper).toBeDefined();
       });
     });
     describe('nested object (not in array)', function() {
       it('shallow render', () => {
-        const wrapper = shallowWithContext(<Form collection={Objects} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={Objects} />);
         expect(wrapper).toBeDefined();
       });
       it('define one field', () => {
-        const wrapper = shallowWithContext(<Form collection={Objects} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={Objects} />);
         const defaultGroup = wrapper.find('FormGroup').first();
         const fields = defaultGroup.prop('fields');
         expect(fields).toHaveLength(1); // addresses field
@@ -201,7 +201,7 @@ describe('vulcan-forms/components', function() {
         return fields;
       };
       const getFirstField = () => {
-        const wrapper = shallowWithContext(<Form collection={Objects} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={Objects} />);
         const fields = getFormFields(wrapper);
         return fields[0];
       };
@@ -212,17 +212,17 @@ describe('vulcan-forms/components', function() {
     });
     describe('array of objects', function() {
       it('shallow render', () => {
-        const wrapper = shallowWithContext(<Form collection={ArrayOfObjects} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={ArrayOfObjects} />);
         expect(wrapper).toBeDefined();
       });
       it('render a FormGroup for addresses', function() {
-        const wrapper = shallowWithContext(<Form collection={ArrayOfObjects} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={ArrayOfObjects} />);
         const formGroup = wrapper.find('FormGroup').find({ name: 'addresses' });
         expect(formGroup).toBeDefined();
         expect(formGroup).toHaveLength(1);
       });
       it('passes down the array child fields', function() {
-        const wrapper = shallowWithContext(<Form collection={ArrayOfObjects} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={ArrayOfObjects} />);
         const formGroup = getArrayFormGroup(wrapper);
         const fields = getFields(formGroup);
         const arrayField = fields[0];
@@ -232,11 +232,11 @@ describe('vulcan-forms/components', function() {
     });
     describe('array with custom children inputs (e.g array of url)', function() {
       it('shallow render', function() {
-        const wrapper = shallowWithContext(<Form collection={ArrayOfUrls} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={ArrayOfUrls} />);
         expect(wrapper).toBeDefined();
       });
       it('passes down the array item custom input', () => {
-        const wrapper = shallowWithContext(<Form collection={ArrayOfUrls} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={ArrayOfUrls} />);
         const formGroup = getArrayFormGroup(wrapper);
         const fields = getFields(formGroup);
         const arrayField = fields[0];
@@ -245,12 +245,12 @@ describe('vulcan-forms/components', function() {
     });
     describe('array of objects with custom children inputs', function() {
       it('shallow render', function() {
-        const wrapper = shallowWithContext(<Form collection={ArrayOfCustomObjects} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={ArrayOfCustomObjects} />);
         expect(wrapper).toBeDefined();
       });
       // TODO: does not work, schema_utils needs an update
       it.skip('passes down the custom input', function() {
-        const wrapper = shallowWithContext(<Form collection={ArrayOfCustomObjects} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={ArrayOfCustomObjects} />);
         const formGroup = getArrayFormGroup(wrapper);
         const fields = getFields(formGroup);
         const arrayField = fields[0];
@@ -259,11 +259,11 @@ describe('vulcan-forms/components', function() {
     });
     describe('array with a fully custom input (array itself and children)', function() {
       it('shallow render', function() {
-        const wrapper = shallowWithContext(<Form collection={ArrayFullCustom} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={ArrayFullCustom} />);
         expect(wrapper).toBeDefined();
       });
       it('passes down the custom input', function() {
-        const wrapper = shallowWithContext(<Form collection={ArrayFullCustom} />);
+        const wrapper = shallowWithContext(<Form collectionName="" collection={ArrayFullCustom} />);
         const formGroup = getArrayFormGroup(wrapper);
         const fields = getFields(formGroup);
         const arrayField = fields[0];
@@ -390,7 +390,7 @@ describe('vulcan-forms/components', function() {
       const wrapper = shallow(<Components.FormNestedObject {...defaultProps} currentValues={{}} />);
       expect(wrapper).toBeDefined();
     });
-    it.skip('render a form for the object', function() {
+    it.skip('render a Form collectionName="" for the object', function() {
       // eslint-disable-next-line no-unused-vars
       const wrapper = shallow(<Components.FormNestedObject {...defaultProps} currentValues={{}} />);
       expect(false).toBe(true);
