+++ packages/vulcan-forms/test/index.js
@@ -1,5 +1,3 @@
 import './schema_utils.test.js';
 import './components.test.js';
 import './mergeWithComponents.test.js';
-
-import './fieldUtils.test.js';
